scenario tx file rename
