+++ framework/scenario/src/facade/world_tx.rs
@@ -1,11 +1,11 @@
 #![allow(unused)] // TEMP
 
-mod scenario_env;
-mod scenario_env_deploy;
-mod scenario_env_exec;
-mod scenario_env_query;
+mod scenario_exec_call;
+mod scenario_exec_deploy;
+mod scenario_query_call;
 mod scenario_rh_impl;
+mod scenario_tx_env;
 
-pub use scenario_env::{ScenarioTxEnv, ScenarioTxEnvData, ScenarioTxRun};
-pub use scenario_env_exec::ScenarioEnvExec;
-pub use scenario_env_query::ScenarioEnvQuery;
+pub use scenario_exec_call::ScenarioEnvExec;
+pub use scenario_query_call::ScenarioEnvQuery;
+pub use scenario_tx_env::{ScenarioTxEnv, ScenarioTxEnvData, ScenarioTxRun};
